requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
strawberry-graphql[fastapi]==0.217.1
sqlalchemy==2.0.25
python-dateutil==2.8.2
